delete some missing folder and delete commented part in python files
