+++ catkin_ws/src/baseline_navi/src/apriltags_to_goalpoint.py
@@ -9,6 +9,7 @@ import tf
 from baseline_navi.msg import TaskStage
 from navigation_controller.srv import command
 from baseline_navi.srv import StageChange
+from baseline_navi.srv import Stage_Totag, Stage_TotagResponse
 import time
 import math
 
@@ -26,10 +27,17 @@ class ApriltagsToGoalPoint(object):
         # Setup the publisher and subscriber
         self.sub_tag = rospy.Subscriber("tag_detections", AprilTagDetectionArray, self.tagCallback)
         self.stage_sub = rospy.Subscriber('baseline_navi/current_stage', TaskStage, self.stage_cb, queue_size = 1)
-        rospy.wait_for_service('baseline_navi/stage_request')
-        self.stage_service = rospy.ServiceProxy('baseline_navi/stage_request', StageChange)
+        self.motion_stage_srv = rospy.Service('baseline_navi/stage_totag', Stage_Totag, self.totag_stage_service_cb)
+        # rospy.wait_for_service('baseline_navi/stage_request')
+        # self.stage_service = rospy.ServiceProxy('baseline_navi/stage_request', StageChange)
         rospy.wait_for_service('pos_cmd')
         self.goal_service = rospy.ServiceProxy('pos_cmd', command)
+
+    def totag_stage_service_cb(self, totag_request):
+        self.stage = 1
+        while self.stage:
+            time.sleep(0.5)
+        return Stage_TotagResponse(True)
 
     def stage_cb(self, stage_msg):
         tmp = stage_msg.current_stage
@@ -69,21 +77,20 @@ class ApriltagsToGoalPoint(object):
         return goal_achived
 
     def tagCallback(self, msg_tags):
-        if self.stage == 2 :
+        if self.stage == 1 :
             goal_pose = PoseStamped()
             self.msg_tags = msg_tags
             self.msg_received = True
-
         # added goal point pub code
             for tag in self.msg_tags.detections:
                 if tag.id[0] == 1:
-                    self.tf_broadcast.sendTransform((0.0, 0.0, 0.6),
+                    self.tf_broadcast.sendTransform((0.0, -0.0, 0.5),
                                       np.array([-0.5, 0.5, 0.5, 0.5]),
                                       rospy.Time.now(),
                                       'goal',
                                       'tag_1')
 
-                    transform_goal = self.tf_buffer.lookup_transform('vslam_map',
+                    transform_goal = self.tf_buffer.lookup_transform('odom',
                                        'goal', #source frame
                                        rospy.Time(0), #get the tf at first available time
                                        rospy.Duration(1.0))
@@ -105,25 +112,20 @@ class ApriltagsToGoalPoint(object):
                     self.last_msg = transform_goal.transform.translation.x
 
                     if self.adjust_counter < 2:
+                        print(transform_goal.transform.translation.x)
+                        print(transform_goal.transform.translation.y)
                         goal_weight = (self.adjust_counter + 2) * 0.25
+                        print(goal_weight)
                         goal_achived = self.send_goal(transform_goal.transform.translation.x * goal_weight,
                             transform_goal.transform.translation.y * goal_weight, y)
                     else:
                         goal_achived = self.send_goal(transform_goal.transform.translation.x, transform_goal.transform.translation.y, y)
+                        print("start turn around")
+                        self.stage = 0
 
                     if goal_achived:
                         if self.adjust_counter < 2:
                             self.adjust_counter += 1
-                        else:
-                            try:
-                                resp = self.stage_service(3, "apriltag_to_goal")
-                                if resp.success :
-                                    print("success change stage to 3")
-                                    self.stage = 3
-
-                            except rospy.ServiceException, e:
-                                print("Service call failed: %s", e)
-
                     time.sleep(1)
 
 if __name__ == '__main__':
