--- catkin_ws/src/baseline_navi_demo/CMakeLists.txt
@@ -1,59 +1,0 @@
-cmake_minimum_required(VERSION 3.0.2)
-project(baseline_navi_demo)
-
-## Compile as C++11, supported in ROS Kinetic and newer
-# add_compile_options(-std=c++11)
-
-## Find catkin macros and libraries
-## if COMPONENTS list like find_package(catkin REQUIRED COMPONENTS xyz)
-## is used, also find other catkin packages
-find_package(catkin REQUIRED COMPONENTS
-   rospy
-   std_msgs
-   message_generation
-)
-
-## Generate actions in the 'action' folder
-# add_action_files(
-#   FILES
-#   Action1.action
-#   Action2.action
-# )
-
-## Generate added messages and services with any dependencies listed here
-# generate_messages(
-#   DEPENDENCIES
-#   std_msgs  # Or other packages containing msgs
-# )
-
-
-add_message_files(
-  FILES
-  TaskStage.msg
-)
-
-add_service_files(
-  FILES
-  StageChange.srv
-  Stage_Grasp.srv
-  Stage_Totag.srv
-)
-
-generate_messages(
-  DEPENDENCIES
-  std_msgs
-)
-
-catkin_package(
-#  INCLUDE_DIRS include
-#  LIBRARIES baseline_navi
-CATKIN_DEPENDS
-    message_runtime
-)
-
-## Specify additional locations of header files
-## Your package locations should be listed before other locations
-include_directories(
-# include
-# ${catkin_INCLUDE_DIRS}
-)
