--- catkin_ws/src/baseline_navi_demo/msg/TaskStage.msg
@@ -1,3 +1,0 @@
-uint32 current_stage
-
-string task_state
