--- catkin_ws/src/baseline_navi_demo/srv/StageChange.srv
@@ -1,4 +1,0 @@
-uint32 desired_stage
-string request_node
----
-bool success
