--- catkin_ws/src/baseline_navi_demo/srv/Stage_Grasp.srv
@@ -1,3 +1,0 @@
-uint32 request
----
-bool success
