--- catkin_ws/src/baseline_navi_demo/srv/Stage_Totag.srv
@@ -1,3 +1,0 @@
-bool request
----
-bool success
