+++ catkin_ws/src/motion_pkg/src/motion_control.py
@@ -22,9 +22,13 @@ from geometry_msgs.msg import Quaternion, PointStamped
 from pyrobot import Robot
 from tf import TransformListener
 from motion_pkg.srv import Motion,MotionResponse
+from motion_pkg.srv import Grasp_Point, Grasp_PointResponse
 from baseline_navi.srv import StageChange
+from baseline_navi.srv import Stage_Grasp, Stage_GraspResponse
+
 from baseline_navi.msg import TaskStage
-from std_msgs.msg import Int32
+from std_msgs.msg import Int32, Bool
+from navigation_controller.srv import command
 
 BB_SIZE = 5
 MAX_DEPTH = 3.0
@@ -44,63 +48,61 @@ class Motion_(object):
 
     def __init__(self):
                 # TODO - use planning_mode=no_plan, its better
+        # self.robot = Robot(
+        #     "locobot_kobuki",
+        #     arm_config={"use_moveit": True, "moveit_planner": "ESTkConfigDefault"},
+        # )
         self.robot = Robot(
-            "locobot_kobuki",
+            "locobot",
             arm_config={"use_moveit": True, "moveit_planner": "ESTkConfigDefault"},
         )
-        self.pregrasp_height = 0.2
-        self.grasp_height = 0.13
+        self.pregrasp_height = 0.17
+        self.grasp_height = 0.11
         self.default_Q = Quaternion(0.0, 0.0, 0.0, 1.0)
         self.grasp_Q = Quaternion(0.0, 0.707, 0.0, 0.707)
         self.retract_position = list([-1.5, 0.5, 0.3, -0.7, 0.0])
         self.reset_pan = 0.0
         self.reset_tilt = 0.8
         self.n_tries = 5
-        self._sleep_time = 2
+        self._sleep_time = 2.5
         self._transform_listener = TransformListener()
-        self.stage_sub = rospy.Subscriber('baseline_navi/current_stage', TaskStage, self.stage_cb, queue_size = 1)
-        self.motion_srv = rospy.Service('locobot_motion', Motion, self.handle_motion)
-        self.grasp_pub = rospy.Publisher("locobot_motion/grasp_start", Int32, queue_size = 1)
-        rospy.wait_for_service('baseline_navi/stage_request')
-        self.stage_service = rospy.ServiceProxy('baseline_navi/stage_request', StageChange)
+        rospy.wait_for_service('locobot_grasppoint')
+        self.grasppoint_service = rospy.ServiceProxy('locobot_grasppoint', Grasp_Point)
+        self.grasp_stage_srv = rospy.Service('baseline_navi/stage_grasp', Stage_Grasp, self.grasp_stage_service_cb)
+        rospy.wait_for_service('pos_cmd')
+        self.goal_service = rospy.ServiceProxy('pos_cmd', command)
         self.color = ""
-        self.stage = 0
-    def stage_cb(self, stage_msg):
-        tmp = stage_msg.current_stage
-        if tmp == 3:
-            self.adjust_counter = 0
-            rospy.loginfo('Going to placing pose')
-            if self.color == 'red':
-                result = self.set_pose([0.234, -0.335, 0.3], roll=0.0)
-            elif self.color == 'green':
-                result = self.set_pose([0.14, -0.335, 0.3], roll=0.0)
-            elif self.color == 'blue':
-                result = self.set_pose([0.024, -0.335, 0.3], roll=0.0)
-            if not result:
-                return False
-            time.sleep(self._sleep_time)
 
-            rospy.loginfo('Opening gripper')
-            self.robot.gripper.open()
-            rospy.loginfo('Going to placing above pose')
-            result = self.set_pose([0.124, -0.335, 0.25], roll=0.0)
-            if not result:
-                return False
+    def loop_service_confirm(self):
+        goal_achived = False
+        while not goal_achived:
             try:
-                resp = self.stage_service(4, "grasp_end")
-                if resp.success :
-                    print("success change stage to 4")
-                    self.stage = 0
-
+                resp = self.goal_service(0, 0, 0, 0)
+                goal_achived = resp.run_completed
             except rospy.ServiceException, e:
                 print("Service call failed: %s", e)
-        elif tmp == 1:
-            self.robot.camera.set_tilt(self.reset_tilt)
-            while self.robot.camera.get_tilt() < self.reset_tilt:
-                print(self.robot.camera.get_tilt())
-                continue
-            self.grasp_pub.publish(1)
- 
+
+        return goal_achived
+
+    def pos_cmd_request(self, request_type, request_x, request_y, request_theta):
+        try:
+            resp = self.goal_service(request_type, request_x, request_y, request_theta)
+        except rospy.ServiceException, e:
+            print("Service call failed: %s", e)
+
+    def send_goal(self, goal_x, goal_y, goal_theta):
+        self.pos_cmd_request(2, goal_x, goal_y, 0)
+        goal_achived = self.loop_service_confirm()
+
+        if not goal_achived:
+            print("Fail at type 2 VSLAM Navigation")
+            return False
+
+        self.pos_cmd_request(1, 0, 0, goal_theta)
+        goal_achive = self.loop_service_confirm()
+
+        return goal_achived
+
     def reset(self):
         """ 
         Resets the arm to it's retract position.
@@ -128,32 +130,27 @@ class Motion_(object):
         """
 
         pregrasp_position = [grasp_pose[0], grasp_pose[1], self.pregrasp_height]
-        # pregrasp_pose = Pose(Point(*pregrasp_position), self.default_Q)
         grasp_angle = self.get_grasp_angle(grasp_pose)
         grasp_position = [grasp_pose[0], grasp_pose[1], self.grasp_height]
-        # grasp_pose = Pose(Point(*grasp_position), self.grasp_Q)
 
         rospy.loginfo("Going to pre-grasp pose:\n\n {} \n".format(pregrasp_position))
         result = self.set_pose(pregrasp_position, roll=grasp_angle)
-        if not result:
-            return False
         time.sleep(self._sleep_time)
 
         rospy.loginfo("Going to grasp pose:\n\n {} \n".format(grasp_position))
         result = self.set_pose(grasp_position, roll=grasp_angle)
-        if not result:
-            return False
         time.sleep(self._sleep_time)
 
         rospy.loginfo("Closing gripper")
         self.robot.gripper.close()
-        time.sleep(self._sleep_time)
+        time.sleep(1)
 
         rospy.loginfo("Going to pre-grasp pose")
         result = self.set_pose(pregrasp_position, roll=grasp_angle)
         if not result:
             return False
-        time.sleep(self._sleep_time)
+        time.sleep(1)
+        print("end of grasp func")
 
     def set_pose(self, position, pitch=DEFAULT_PITCH, roll=0.0):
         """ 
@@ -176,6 +173,7 @@ class Motion_(object):
                 position=position, pitch=pitch, roll=roll, plan=False, numerical=False
             )
             if success == 1:
+                print("set_pose success")
                 break
         return success
 
@@ -213,7 +211,7 @@ class Motion_(object):
         """
         self.exit()
 
-    def handle_motion(self, req):
+    def handle_grasp(self, req):
         rospy.loginfo("Grasp attempt x={:.4f},y={:.4f},theta={:.4f}".format(req.x,req.y,req.theta))
         self.color = req.color
         try:
@@ -227,8 +225,43 @@ class Motion_(object):
         self.robot.camera.set_tilt(0.0)
         self.grasp(grasp_pose)
 
+        #robotics arm placing
+        rospy.loginfo('Going to placing pose')
+        print(self.color)
+        if self.color == 'red':
+            result = self.set_pose([0.234, -0.235, 0.3], roll=0.0)
+            print("red")
+        elif self.color == 'green':
+            result = self.set_pose([0.174, -0.235, 0.3], roll=0.0)
+            print("green")
+        elif self.color == 'blue':
+            result = self.set_pose([0.114, -0.235, 0.3], roll=0.0)
+            print("blue")
+        if not result:
+            print("False")
+            return False
+        time.sleep(self._sleep_time)
+
         return MotionResponse("ok")
 
+    def grasp_stage_service_cb(self, grasp_request):
+        if grasp_request.request == 1:
+            self.reset()
+            self.send_goal(0, 0, -1.57)
+            print("call grasp.py")
+            response = self.grasppoint_service(True)
+            print("get point")
+            self.handle_grasp(response)
+            self.send_goal(0, 0, 0)
+            return Stage_GraspResponse(True)
+        elif grasp_request.request == 3:
+            rospy.loginfo('Opening gripper')
+            self.robot.gripper.open()
+            rospy.loginfo('Going to placing above pose')
+            self.set_pose([0.15,0.0,0.3],roll=0.0)
+            rospy.loginfo('back to original point')
+            self.send_goal(0, 0, 0)
+            return Stage_GraspResponse(True)
  
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="Process args for grasper")
@@ -239,8 +272,7 @@ if __name__ == "__main__":
                         dest='display_grasp', action='store_false')
     parser.set_defaults(no_visualize=True)
 
-    args = parser.parse_args()
+    args, unknown = parser.parse_known_args()
     rospy.init_node('locobot_motion_server')
     motion = Motion_()
-    motion.reset()
     rospy.spin()
