--- catkin_ws/src/motion_pkg_demo/srv/Grasp_Point.srv
@@ -1,6 +1,0 @@
-bool request
----
-float64 x
-float64 y
-float64 theta
-string color
