--- catkin_ws/src/motion_pkg_demo/srv/Motion.srv
@@ -1,6 +1,0 @@
-float64 x
-float64 y
-float64 theta
-string color
----
-string status
